Refactor to json_utils
