+++ gramps/gen/lib/json_utils.py
@@ -51,6 +51,15 @@ def convert_object_to_state(obj):
     return obj.get_object_state()
 
 
+def string_to_data(string):
+    return DataDict(orjson.loads(string))
+
+def string_to_dict(string):
+    return orjson.loads(string)
+
+def dict_to_string(dict):
+    return orjson.dumps(dict)
+
 def object_to_data(obj):
     """
     Convert any Gramps lib object into its dict representation.
@@ -89,12 +98,75 @@ def object_to_string(obj: object) -> str | bytes:
     """
     return orjson.dumps(object_to_data(obj))
 
+def data_to_string(data):
+    return orjson.dumps(data)
+
 
 def string_to_object(string: str | bytes):
     """
     Convert a JSON string/bytes into a Gramps lib object.
     """
     return data_to_object(orjson.loads(string))
+
+
+class DataDict(dict):
+    """
+    A wrapper around a data dict that also provides an
+    object interface.
+    """
+
+    def __init__(self, data=None):
+        """
+        Wrap a data dict (raw data) or object
+        with an attribute API. If data is an
+        object, we use it to get the attributes.
+        """
+        if isinstance(data, dict):
+            super().__init__(data)
+        else:
+            super().__init__()
+            if data:
+                self["_object"] = data
+
+    def __str__(self):
+        if "_object" not in self:
+            self["_object"] = from_dict(self)
+        return str(self["_object"])
+
+    def __getattr__(self, key):
+        if key.startswith("_"):
+            raise AttributeError(
+                "this method cannot be used to access hidden attributes"
+            )
+
+        if "_object" in self:
+            return getattr(self["_object"], key)
+        elif key in self:
+            value = self[key]
+            if isinstance(value, dict):
+                return DataDict(value)
+            elif isinstance(value, list):
+                return DataList(value)
+            else:
+                return value
+        else:
+            self["_object"] = from_dict(self)
+            return getattr(self["_object"], key)
+
+
+class DataList(list):
+    """
+    A wrapper around a data list.
+    """
+
+    def __getitem__(self, position):
+        value = super().__getitem__(position)
+        if isinstance(value, dict):
+            return DataDict(value)
+        elif isinstance(value, list):
+            return DataList(value)
+        else:
+            return value
 
 
 from_json = string_to_object
