+++ gramps/gen/lib/serialize.py
@@ -39,74 +39,19 @@ LOG = logging.getLogger(".serialize")
 # Gramps modules
 #
 # ------------------------------------------------------------------------
-from .serialize_utils import (
+from .json_utils import (
     from_dict,
     to_dict,
     to_json,
     from_json,
-    json_loads,
-    json_dumps,
+    dict_to_string,
+    data_to_object,
+    string_to_object,
+    string_to_data,
+    string_to_dict,
+    object_to_string,
+    data_to_string,
 )
-
-
-class DataDict(dict):
-    """
-    A wrapper around a data dict that also provides an
-    object interface.
-    """
-
-    def __init__(self, data=None):
-        """
-        Wrap a data dict (raw data) or object
-        with an attribute API. If data is an
-        object, we use it to get the attributes.
-        """
-        if isinstance(data, dict):
-            super().__init__(data)
-        else:
-            super().__init__()
-            if data:
-                self["_object"] = data
-
-    def __str__(self):
-        if "_object" not in self:
-            self["_object"] = from_dict(self)
-        return str(self["_object"])
-
-    def __getattr__(self, key):
-        if key.startswith("_"):
-            raise AttributeError(
-                "this method cannot be used to access hidden attributes"
-            )
-
-        if "_object" in self:
-            return getattr(self["_object"], key)
-        elif key in self:
-            value = self[key]
-            if isinstance(value, dict):
-                return DataDict(value)
-            elif isinstance(value, list):
-                return DataList(value)
-            else:
-                return value
-        else:
-            self["_object"] = from_dict(self)
-            return getattr(self["_object"], key)
-
-
-class DataList(list):
-    """
-    A wrapper around a data list.
-    """
-
-    def __getitem__(self, position):
-        value = super().__getitem__(position)
-        if isinstance(value, dict):
-            return DataDict(value)
-        elif isinstance(value, list):
-            return DataList(value)
-        else:
-            return value
 
 
 class BlobSerializer:
@@ -171,22 +116,22 @@ class JSONSerializer:
             "json, data_to_object: {'_class': %r, ...}",
             data["_class"] if (data and "_class" in data) else data,
         )
-        return from_dict(data)
+        return data_to_object(data)
 
     @staticmethod
     def string_to_object(obj_class, string):
         LOG.debug("json, string_to_object: %r...", string[:65])
-        return from_json(string)
+        return string_to_object(string)
 
     @staticmethod
     def string_to_data(string):
         LOG.debug("json, string_to_data: %r...", string[:65])
-        return DataDict(json_loads(string))
+        return string_to_data(string)
 
     @staticmethod
     def object_to_string(obj):
         LOG.debug("json, object_to_string: %s...", obj)
-        return to_json(obj)
+        return object_to_string(obj)
 
     @staticmethod
     def data_to_string(data):
@@ -194,22 +139,18 @@ class JSONSerializer:
             "json, data_to_string: {'_class': %r, ...}",
             data["_class"] if (data and "_class" in data) else data,
         )
-        return json_dumps(data)
+        return data_to_string(data)
 
     @staticmethod
     def metadata_to_object(string):
-        doc = json.loads(string)
+        doc = string_to_dict(string)
         type_name = doc["type"]
-        if type_name in ("int", "str", "list"):
-            return doc["value"]
-        elif type_name == "set":
+        if type_name == "set":
             return set(doc["value"])
         elif type_name == "tuple":
             return tuple(doc["value"])
-        elif type_name == "dict":
-            return doc["value"]
         elif type_name == "Researcher":
-            return from_dict(doc["value"])
+            return data_to_object(doc["value"])
         else:
             return doc["value"]
 
@@ -219,11 +160,10 @@ class JSONSerializer:
         if type_name in ("set", "tuple"):
             value = list(value)
         elif type_name == "Researcher":
-            value = to_dict(value)
-        elif type_name not in ("int", "str", "list"):
-            value = json.loads(to_json(value))
+            value = object_to_data(value)
+
         data = {
             "type": type_name,
             "value": value,
         }
-        return json.dumps(data)
+        return dict_to_string(data)
